http://issues.apache.org/activemq/browse/AMQCPP-16
CMS API Updates

git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/activemq-cpp/trunk/activemq-cpp@479711 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/connector/openwire/commands/ActiveMQBytesMessage.h
@@ -253,7 +253,8 @@ namespace commands{
          * @throws CMSException
          */
         virtual void setBodyBytes( 
-            const unsigned char* buffer, const unsigned long numBytes ) 
+            const unsigned char* buffer, 
+            const unsigned long long numBytes ) 
                 throw( cms::CMSException );
             
         /**
@@ -269,7 +270,7 @@ namespace commands{
          * Returns the number of bytes contained in the body of this message.
          * @return number of bytes.
          */
-        virtual unsigned long getBodyLength(void) const;
+        virtual unsigned long long getBodyLength(void) const;
 
     };
